# Conflicts:
#   server/src/main/java/com/zzuiksa/server/domain/gifticon/data/response/GifticonPreviewDto.java

+++ server/src/main/java/com/zzuiksa/server/domain/place/data/MemberPlaceDto.java
@@ -26,10 +26,10 @@ public class MemberPlaceDto {
     private String name;
 
     @NotNull
-    private Float lat;
+    private Double lat;
 
     @NotNull
-    private Float lng;
+    private Double lng;
 
     public static MemberPlaceDto from(Place place) {
         return MemberPlaceDto.builder()
